Add xgit_get_commits() and supporting functions.
This has to deal with a few special cases, like if one object or the other is
empty. If the new object is empty, that is not so hard, since no commits are
returned. If the old object is empty, however, it means that a branch is being
created.

When a branch is created, we are only interested in the commits (if any) that it
is adding to the repository. In other words, if a new branch is created at a
commit that already existed within the repository, then there are no commit
restrictions which apply other than the branch name. If the branch does bring in
new commits, then each of those commits must be allowed in order for the branch
to be permitted.

However, things get tricky when you realize that care must be taken in order to
check only the correct commits. Check too many commits and a commit which was
performed by a more privileged user could be mistakenly included in the "can I
commit this" query from a less privileged user and fail, even though the less
privileged user was not actually trying to push anything they shouldn't be able
to. To few commits checked and the repository could end up accepting commits it
shouldn't have.

The correct thing to do is to check all the ancestors of the new value of the
pushed ref, except for those which are reachable from already-existing branches
in the repository. This means that if the branch introduces new commits, those
commits won't be reachable by any other ref in the repository, and so will be
returned by 'git rev-list' to be checked.

Signed-off-by: Dan Hackney <chrono325@gmail.com>