Read marks and return them from fastImport().
Returns the marks created during the execution of git-fast-import(1) as an array
of (markid => SHA-1).

Also, in order to clean up its own temporary files, fastImport() unlinks the
file specified by the 'export-marks' option. If it is needed after fastImport()
returns, it should be opened with fopen() before calling fastImport().

Signed-off-by: Dan Hackney <chrono325@gmail.com>