Add helper functions assertMarkBlob() and assertObjectStats().
assertMarkBlob() checks for the existence of a mark, that its type is a blob,
and checks the value of the blob.

assertObjectStats() Looks at the stderr of fast-import and checks for a specific
number of created objects, whether they are blobs, trees, commits, tags, or
marks. It can check for a specific number of objects created, duplicated, or
both.

Signed-off-by: Dan Hackney <dan@haxney.org>