Check if commit command has a 'merge' field before adding it.
If the commit command did not have a 'merge' field, then a NULL variable was
supplied to the foreach loop. It would be okay if it were an empty array, since
the loop would be run zero times. Check that the array is not empty before
trying to iterate over it.

Signed-off-by: Dan Hackney <dan@haxney.org>