Make readMarksFile()'s return array integer- rather than string-keyed.
The marks file places a colon (":") before the mark id number. Strip the colon
and use an integer as the key for the array. This removes the need for users of
the marks array to add colons to the mark numbers they use. It is also
consistent with the input to GitFastImport::fastImport(), as the 'mark' options
all take integers rather than strings.

Signed-off-by: Dan Hackney <dan@haxney.org>