More efficiently handle time conversion for author and committer date.
Use the return value from checkAndGet($author, 'when') rather than discarding
it. Prevents having to look the value up again.

Also, don't pass the value of $required to the checkAndGet() for 'when', since
if the author field is not required and is not present, then the function will
return before getting there. If the author option is present, then all of the
required parts of the author array should be present as well.

Signed-off-by: Dan Hackney <dan@haxney.org>