Creat GitFastImport::MARK_ID_MAX.
For some reason (related to implementation details of git-fast-import), marks
can only go up to 2^20 - 1 (= 1,048,575) before the behavior becomes
unspecified. If a higher mark is given, the mark id written to the
"--export-marks" file is not equal to the mark id given, and bears no clear
relationship to the input number.

Any function using marks should avoid using marks above this limit, or the
result will be unspecified.

Signed-off-by: Dan Hackney <dan@haxney.org>